backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
bcrypt
PyJWT
